QSDK:PROTOCOL:HOSTAPD: RADIUS client vulnerability bug fixes
1) Drop pending request only when accepting the response
The case of an invalid authenticator in a RADIUS response could imply
that the response is not from the correct RADIUS server and as such,
such a response should be discarded without changing internal state for
the pending request. The case of an unknown response (RADIUS_RX_UNKNOWN)
is somewhat more complex since it could have been indicated before
validating the authenticator. In any case, it seems better to change the
state for the pending request only when we have fully accepted the
response.

Allowing the internal state of pending RADIUS request to change based on
responses that are not fully validation could have allow at least a
theoretical DoS attack if an attacker were to have means for injecting
RADIUS messages to the network using the IP address of the real RADIUS
server and being able to do so more quickly than the real server and
with the matching identifier from the request header (i.e., either by
flooding 256 responses quickly or by having means to capture the RADIUS
request). These should not really be realistic options in a properly
protected deployment, but nevertheless it is good to be more careful in
processing RADIUS responses.

Remove a pending RADIUS request from the internal list only when having
fully accepted a matching RADIUS response, i.e., after one of the
registered handlers has confirmed that the authenticator is valid and
processing of the response has succeeded.

PATCH: r61-00-QSDK-PROTOCOL-HOSTAPD-Drop-pending-request-when-accepting-response.patch
https://w1.fi/cgit/hostap/commit/?id=726432d7622cc0088ac353d073b59628b590ea44

2) Fix pending request dropping
A recent change to this moved the place where the processed RADIUS request was removed
from the pending list to happen after the message handler had been called.
This did not take into account possibility of the handler adding a new
pending request in the list and the prev_req pointer not necessarily pointing
to the correct entry anymore. As such, some of the pending requests could
have been lost and that would result in not being able to process responses
to those requests and also, to a memory leak.

Fix this by determining prev_req at the point when the pending request
is being removed, i.e., after the handler function has already added a
new entry.

PATCH: r61-01-QSDK-PROTOCOL-HOSTAPD-Fix-pending-request-dropping.patch
https://w1.fi/cgit/hostap/commit/?id=339a334551ca911187cc870f4f97ef08e11db109

Tag: RADIUS
Patch-work: None
Patch-dependency: None

Change-Id: Ic72b40b8a85a804cfbf21621e5c2f8004d7adfb9
Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>
Signed-off-by: Arunpandi Kannan <quic_arunpand@quicinc.com>