QSDK:PROTOCOL: wpa_supplicant: Fix deauth and re-association in 5 GHzl and 5 GHzh split phy cases
Currently when the number of associated links is lesser than the
usual configuration, then for every 5 seconds existing scan entries
will be fetched from driver and will try to deauthenticate
and re-associate again. If the frequency count of all the links
configured is equal to the number of valid links, then it is
considered that all configured links are associated and there is
no need to run the background scan for re-association.

For split-phy cases having 5 GHzl and 5 GHzh, the frequency count is
taken the same instead of having it separately which results in
executing background scan and in turn results in deauthentication
followed by re-association if a link is removed and added back in
5 GHzl and 5 GHzh.

Fix by checking if the frequency count is lesser than links count
instead of checking whether they are equal.

Fixes: fc72724db0fc ("[WAR] wpa_supplicant: Add changes to reassociate
if")

Change-Id: I2706e34fdff86d40936b9b6bb927e28cb4f02ff3
Signed-off-by: Dhanavandhana Kannan <quic_dhanavan1@quicinc.com>
(cherry picked from commit c0654db268452662e15dc72bb00445ce8c423c6a)