QSDK:PROTOCOL:HOSTAPD:hostapd DOS attack Vulnerability fixes SAE.
When processing the initial SAE authentication frame (i.e., the SAE commit message)
hostapd stores the received password identifier in its internal state.
This was done even if the password identifier was not known. The stored password
identifier is then used to verify that the following SAE messages are using
the same password identifier. This could result in an attacker being able to
cause denial of service for a legitimate STA by injecting a frame on behalf of
that STA with a password identifier that the STA is not using. This would cause
hostapd to discard SAE authentication frames from the legitimate STA as long as
the internal state for that STA is maintained. If the STA is not explicit
deauthenticating, this could result in maintaining state indefinitely

PATCH: r62-00-QSDK-PROTOCOL-HOSTAPD-Reject-unexpected-password-identifier-sae.patch
https://w1.fi/cgit/hostap/commit/?id=8a33510550cb29da62df00b87623c7cf6ae133aa
SAE: Reject unexpected password identifier in commit message parser

PATCH: r62-01-QSDK-PROTOCOL-HOSTAPD-More-robust-password-identifier-check-in-AP.patch
https://w1.fi/cgit/hostap/commit/?id=28ebad01e8d35827672ff562d4a23947dff23229
SAE: More robust password identifier checks for AP mode

PATCH: r62-02-QSDK-PROTOCOL-HOSTAPD-Remove-STA-entries-if-associat.patch
https://w1.fi/cgit/hostap/commit/?id=07ee3e8b0eb4026a0affb209b3d44ba232e532e9
Remove STA entries if association is not completed in 60 seconds

PATCH: r62-03-QSDK-PROTOCOL-HOSTAPD-Do-not-use-the-wpa_passphrase-if-sae-is-used.patch
https://w1.fi/cgit/hostap/commit/?id=90a3b4a91a5dff29b8e8431983aacfc7aad52381
SAE: Do not use the wpa_passphrase if SAE password identifier was used

Patch-dependency: None
Patch-work: None
Tag: SAE

Change-Id: I25cbee885db7624ec1d56020277e76cb86b455bd
Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>
Signed-off-by: Arunpandi Kannan <quic_arunpand@quicinc.com>