UPSTREAM:PROTOCOL: Fix legacy client association across different radios
When a legacy station associates with one radio of an MLD AP and leaves
without disassociation or deauthentication, hostapd maintains the
station context until the maximum inactivity timer expires. This causes
a problem when the station tries to re-associate as a legacy station on
a different radio of the same MLD AP. After EAPOL, before setting the
station as authorized on new link, if the station information is present
on the partner link, the station entry is removed as part of pruning
after sending disassociation on the first radio where it was originally
associated. With MLO, all links of the same MLD use the same netdev, so
pruning the legacy station removes the kernel entry completely, leading
to repeated association and pruning until the station context is removed
due to maximum inactivity on the first link the station was associated.

Improve this by skipping pruning if the station is authorized and tries
to associate in partner links of the same MLD. This ensures that the
station can successfully associate as a legacy station on a different
radio without encountering association failures. This change prevents
the removal of the kernel entry for authorized stations, allowing
seamless re-association across different radios within the same MLD AP.

Patch-dependency: None

Patch-work: None

Tag: ASSOC

Change-Id: I05601757f12d33995e24848fc87a624faad0cfc6
Signed-off-by: Hari Naraayana Desikan Kannan <quic_hnaraaya@quicinc.com>