QSDK:CP Add max_oper_chwidth 0 for bandwidth EHT20 and EHT40
disabled_vht is set to 1 for the 2G mesh in both EHT20 and EHT40
bandwidths. This configuration is not populating the VHT capabilities
in the association, which is negatively impacting the KPI in the 2G mesh
instead need to add max_oper_chwidth 0 for EHT20 and EHT40 bandwidths.

Patch-dependency: None
Patch-work: None

Change-Id: If4995961fedc5b32c26c7bce2dcbd3caf11bc13e
Signed-off-by: Harish Rachakonda <quic_rachakon@quicinc.com>