UPSTREAM:HOSTAPD: Update conf->op_class in hostapd_change_config_freq
When there are 16 AP MLDs and a channel switch happens on the 6 GHz
link, if a station is connected to the AP on two links (5 GHz and
6 GHz), the station disconnects from the AP and reconnects only on
the 5 GHz link. This disconnection occurs when the AP is operating on
320 MHz and the target channel's bandwidth is 80 MHz.

On the station side, the EHT operation information present bit seems
to be set, which should not happen when the new target channel
bandwidth is 80 MHz. The EHT operation information present bit should
be set only for 320 MHz or for puncturing cases.

On the AP side, the first few beacons on the new target channel
(operating on 80 MHz) have the EHT operation information present
bit set and the ccfs0 and ccfs1 with incorrect values. This is because,
in hostapd_eid_eht_operation, the ch_width is calculated based on the
conf->opclass. During the channel switch, hostapd_change_config_freq
does not assign the target channel's op_class to conf->op_class
during the construction of the beacon. This leads to the width
remaining as 320 MHz (current channel bandwidth), but the seg0 value
is that of an 80 MHz (target channel bandwidth) channel.

To fix this issue, assign conf->op_class in hostapd_change_config_freq.
This will ensure that the ch_width in hostapd_eid_eht_operation is
80 MHz, and thus the EHT operation element present bit will not be set,
and the control field will not be present in the AP's beacon.

Feature: CHANNEL_SWITCH
Patch-work: Update conf->op_class in hostapd_change_config_freq
Patch-dependency: None

Change-Id: Ief2dc6292298f92a90076daf519c47777872e2eb
Signed-off-by: Hariharan Basuthkar <quic_hbasuthk@quicinc.com>