UPSTREAM:PROTOCOL:hostapd: Fix hostapd Crashes Due to interface Setup Failures
Hostapd crashes when it fails to set up the interface in
hostapd_add_iface due to invalid channel information, followed by
toggling the WiFi down and up.

Currently, when the interface setup fails, it only deinitializes
the driver. The BSS link is not deinitialized, which causes the
hostapd crash.

Add the fix to deinitialize the BSS link if the interface setup
fails.

Patch-dependency: None

Change-Id: I6b7a80ed5c37a80d1d80029917c3391daf4985dd
Signed-off-by: Ramasamy Kaliappan <quic_rkaliapp@quicinc.com>