QSDK:HOSTAPD : Fix ACS on 240 MHz channel
When a AP is configured with channel as 'auto' and bandwidth
as HT 320 with puncturing pattern 0xF000 onn 5 GHz, the AP fails
to come up after ACS.

The structure bw_320_1 did not have the 5 GHz 240 Mhz entry.
Due to this, the 5 GHz 240 MHz channels is not available for
ACS to choose this channel.

To fix this issue, add  5 GHz 240 Mhz channel entry in bw_320_1, and
make necessary changes to ACS support APIs to handle it. Also,
assign (*ideal_chan)->punct_bitmap to the bitmap in the iface->conf

Feature: ACS
Patch-dependency: None
Fixes: Id76ab86c517 ('hostapd: Add ACS and DFS support for 5GHz BW240MHZ')

Change-Id: I2e14971357a9a36e7a23c837bb4cda0b2d2d62d8
Signed-off-by: Hariharan Basuthkar <quic_hbasuthk@quicinc.com>