UPSTREAM:PROTOCOL:wpa_supplicant: Fix removed BSS not found during successive link removals
Currently when the first primary link is removed via dynamic link
removal, wpa_flush_bss_by_age() eliminates inactive BSS from the
supplicant scan results after a certain period. When the second
primary link is subsequently removed, the supplicant sends a
deauthentication message because the BSS of the first removed primary
link is no longer found in the scan results. This occurs when the
requested links during association are checked for their presence in
the scan results within the wpa_sm_set_ml_info() function during link
removal.

To fix this, ensure that if a removed BSS is not present in the scan
results due to the flush, the BSS not found failure is skipped by
verifying that the removed BSS does not belong to valid links.

Patch-dependency: r32-006-Hostapd-Handle-EVENT_RX_MGMT-event-for-link-
                  reconfig.patch
Patch-work: None

Change-Id: I1b51379190dd2514e9ed67065f3727832e898e16
Signed-off-by: Dhanavandhana Kannan <quic_dhanavan1@quicinc.com>