UPSTREAM:PROTOCOL: Fix updating tbtt_count in EVENT_LINK_REMOVAL
eht_mld_link_removal_inprogress is used to identify
whether the link reconfigure removal is in progress for the link of an MLD AP.

This flag is set when the link reconfigure removal starts and cleared when the
the link reconfigure removal completes. It is not required to be set in the EVENT_LINK_RECONFIG.EVENT_LINK_RECONFIG updates the user space with the link_removal TBTT_COUNT and link_id of the link undergoing removal. The function switch_link_hapd takes the link_id of the link undergoing removal and returns the BSS of the respective link.  However, in some cases, if respective link is removed and the BSS is not found it returns the hapd  which results in updating the wrong hapd's TBTT_COUNT.

Therefore, remove setting eht_mld_link_removal_inprogress in
EVENT_LINK_REMOVAL. Also, add a condition to update the count in EVENT_LINK_RECONFIG only if the eht_mld_link_removal_inprogress is true.

Patch-dependency: r32-005-hostapd-Handle-link-reconfig-event-started-and-compl.patch

Patch-work: None

Fixes: I2383a1e07 (hostapd: Remove Reconfig ML IE as default)

Change-Id: I72b57a2edea1ed0765a5174404542c80c9effe5c
Signed-off-by: Yuvarani V <quic_yuvarani@quicinc.com>
(cherry picked from commit 1492d9f4514b75377b492676bfe5e4c60b3b6a35)