QSDK: PROTOCOL: hostapd: Fix memleak issue in radius server
For each EAP session, a set of variables are allocated memory. Before
going to the successive session, the previous session's memory must be
deallocated. But, currently, memory allocated for accept_attr is not
deallocated before moving the the next session, resulting in memory leak.

Fix this by freeing the allocated memory before allocating memory for
accept_attr in the next session.

Depends-on: None
Fixes: None

Change-Id: Ia5dec4a0d8fb498fcaa4b36da62d9f8d71b9020d
Signed-off-by: Roopni Devanathan <quic_rdevanat@quicinc.com>