QSDK:PROTOCOL:wpa_supplicant: Improvise WAR for max MLD link association
Current implementation will trigger a background scan periodically if
station is associated to lesser number of links than the user configured
links. The lesser number of assoc links is might be due to either one of
the AP MLD is undergoing a CAC/ACS or AP MLD is originally configured
with lesser links. Now, when a link is enabled after CAC/ACS or
dynamically added, then the background scan will trigger a deauth once
the RNR of associated beacon is updated with affiliated AP details.

In certain scenarios, associated beacon update from firmware doesn't
update the entire IE, instead only the beacon_ie_len field alone getting
updated. During this case, though a new link is added in the AP MLD,
station won't re-associate as it couldn't find the RNR elements.

Hence, improvise the WAR logic by triggering a explicit scan in the
existing BSS, if there is a change in the beacon_ie_len field. With
this, post the scan triggered associated beacon is expected to be
updated with the latest RNR element which will contain the newly enabled
link in AP. Once, the RNR is found and the frequency matches with the
configured list in station, trigger deauthentication to make station
associate with max MLD link.

Change-Id: I51ae45fe8e9b2ff3b290d86110ff5a7c98692954
Signed-off-by: Manish Dharanenthiran <quic_mdharane@quicinc.com>