UPSTREAM:PROTOCOL:hostapd: Fix removal of Tx BSS of other MBSSID Set
Currently, if a MLD holds only a non-transmitting BSS(for example,
an MLD with only a 6 GHz link) in a MBSSID set, attempting to
remove the transmitting BSS in another MBSSID set results in a
removal failure, indicating that link reconfiguration is not
applicable for a list size of 1.

When the transmitting BSS is initiated for removal, it also remove
the associated non-transmitting BSS. If there is only one link present
in the non-transmitting BSS within an MLD, the transmitting BSS cannot
be removed because at least one link must be present in the associated
non-transmitting BSS.

With the introduction of MBSSID grouping, the removal of a transmitting
BSS should only consider the non-transmitting BSSs within that specific
MBSSID set to determine if link reconfiguration is permissible. However,
the current implementation checks for non-transmitting BSSs across the
entire interface, rather than the specific MBSSID set, resulting in
removal failures.

The fix involves checking each MBSSID set individually when determining
if link reconfigurtaion is allowed, instead of checking the entire
interface, to avoid this issue.

Fixes: Ie2ba6a8220a7(hostapd: Add Multi-Link Reconfigure link removal
support)

Patch-dependency: r32-002-hostapd-Remove-TESTING_OPTIONS-compile-
		  flag-dependen.patch
Patch-work: None

Change-Id: Iaeaaa834622f6b22640e264b380f0c07a4008ea3
Signed-off-by: Dhanavandhana Kannan <quic_dhanavan1@quicinc.com>