UPSTREAM:CP: Set wds_mld_uid for all link STA
Currently when the primary link is removed via ML reconfig in a WDS
connection and the STA gets disconnected, the VLAN is not deleted,
resulting in a stale entry. This occurs because the wrong interface
name is sent to the driver.

Fix is to set wds_mld_uid flag for all the links to ensure
correct interface is sent to driver.

Patch-dependency: r03-029-hostapd-Enable-WDS-AP-MLO-Support.patch
Patch-work: None

Change-Id: Ief681a992753b69bd5796e5f859345d142592014
Signed-off-by: Harish Rachakonda <quic_rachakon@quicinc.com>