QSDK: WAR:PROTOCOL: Add cases to check for split-phy cases during reassociation
Currently, is_all_link_associated() API will check whether all the
frequency list configured by the user is associated with AP or not.
Association might not have happened because either AP will be in CAC/ACS
or it will get added dynamically later. Here, frequencies configured is
checked and list is saved only for 2 GHz, 5 GHz and 6 GHz individually.
But, it didn't considered about whether underlying hardware has
split-phy or not. Because of this, if user configures both split-phy
frequencies, if either one of the frequency is associated, station won't
try to reconnect even after it starts to beaconing.

Fix this by considering whether split-phy case is present and save the
frequency list accordingly.

Change-Id: I125ca82e30b18a94776686d7d117cbc0729a15ae
Signed-off-by: Manish Dharanenthiran <quic_mdharane@quicinc.com>