UPSTREAM:PROTOCOL: Handle Dynamic link address change for connected  STA
When an already connected Multi-Link Device (MLD) Station (STA) sends a
reassociation request with a different link address, the Access Point (AP)
initiates a Security Association (SA) query to the MLD address it has
on record. However, the STA does not receive the SA query,
leading to a timeout as AP sends out with old ML address.
Consequently, the AP rejects the STA's reassociation request since
the AP will not have station entry.

The STA then restarts the authentication process with the AP,
but this new request also contains a different link address.
As a result, the STA repeatedly comes up new link address and
fails to connect.

If an authentication request from a MLD STA is received with
a SA query timeout, the AP update the new link address from
the authentication request and re-add the STA.
When a previously connected STA returns with a new link address but
retains the old ML address, Hostapd sends out to driver with
new link address, the driver translates the Multi-Link (ML)
address to the link address upon response. In this scenario,
the authentication response will be sent to the old link address,
which the STA might miss. To prevent this, send the authentication
response to the same SA address, ensuring the driver does not perform
the translation and the packet reaches the intended destination.
The AP will then send an authentication reply with the updated link
address for the MLD STA instead of the MLD address.
Add support to update the SA query timeout to all the linked stations
of MLD STA, It helps maintaining integrity during Reassociation.

Patch-work: [PATCH] Hostapd: Handle Dynamic link address change
            for connected  STA
Patch-dependency: None

Change-Id: I4a34ddc006bfd2277b814e37758de1eb8a948751
Co-developed-by: Arunpandi Kannan <quic_arunpand@quicinc.com>
Signed-off-by: Arunpandi Kannan <quic_arunpand@quicinc.com>
Signed-off-by: Karthik M <quic_karm@quicinc.com>