UPSTREAM:PROTOCOL:hostapd: Fix maximum MBSSID group size
The current maximum MBSSID group size is set to 16, but group size
should be between 2 and 8 because when set to 16, the beacon size
would exceed the default size of 1500 due to reduced neighbor
report element that adds the neighbor information all the 16 vaps.

To fix this, the maximum group size should be set to 8,
ensuring that vaps can only be brought up when the group size
is within the 2 to 8 range.

Fixes: I3b1e1d6f2006(hostapd: Add multi MBSSID group support)

Patch-dependency: r45-004-hostapd-Add-configuration-for-Multi-
		  MBSSID.patch

Patch-work: None

Change-Id: Ieb6bb6063a3cea4818afeb79aaa9e0e7f5783f48
Signed-off-by: Dhanavandhana Kannan <dhanavan@qti.qualcomm.com>