QSDK:CP: Fix new interface creation and AID assignment in multi_ap mode
When using multi_ap in both AP and STA modes, an unintended AP/VLAN
interface was being created and the AID was incorrectly assigned.
This was due to an incorrect flag check and the handling of
multi_ap flags.

This patch corrects the logic by:
- Adding proper checks for multi_ap flags to prevent unintended
interface creation.
- Fixing AID assignment by including the multi_ap flag in the
condition used for AID selection.

Patch-work: None
Patch-dependency: None

Change-Id: I41fb9984b028d23baf1e628087da4e44140a5f92
Signed-off-by: Gunaal R <gunar@qti.qualcomm.com>