UPSTREAM:PROTOCOL: hostapd: Add support for Advertised TID-to-Link Mapping
Traffic Identifier (TID) values are used to classify different types of
network traffic, such as voice, video, and data. Link mapping involves
associating these TID values with specific links in a multi-link
operation (MLO) environment. This patch series targets to enable an AP
MLD to advertise Traffic Identifier (TID)-to-Link Mapping (TTLM) in
beacon frames to inform connected devices about the mapping of traffic
identifiers to specific links.

A hostapd_cli command is added to configure the new mapping to be
established, the duration after which the mapping should be applied
(Mapping switch time)and how long is the mapping to be applied
(Expected duration). The AP advertises the mapping through the TTLM
IE in the beacon.

In beacon offload mode, after the TTLM configurations reach firmware,
it converts the Mapping Switch Time(MST) which was supllied in the units
of TU into TSF and sends it back to hostapd. This value is to be used in
the subsequent probe responses and assoc responses. The firmware also
sends events notifying the expiry of MST and Expected Duration(ED).
All associated data structures are to be updated based on these expiry
event to maintain the information about the link mapping that is
scheduled to be advertised or the mapping that is already being
advertised.

Patch-dependency:
r80-001-hostapd-Add-support-to-send-TTLM-request-action.patch
r80-003-hostapd-Add-support-to-handle-TTLM-negotiation-via.patch

Feature: TTLM

Change-Id: Idf875f4cdcab5cc21ee00f5b545ae9ddb62be4cc
Signed-off-by: Amith A <amitajit@qti.qualcomm.com>