UPSTREAM:PROTOCOL: hostapd: Negotiation of TID-To-Link Mapping feature support
Add support for Negotiation of Tid-To-Link-Mapping(TTLM) feature
for AP MLD.

TID-to-Link mapping mechanism allows an AP MLD or non-AP MLD that
performed or performing multi-link setup to determine how uplink(UL)
and downlink(DL) Qos traffic corresponding to TID values between 0
to 7 will be assigned to the setup links for the non-AP MLD.

By default, all the TIDs are mapped to all the setup links for both
UL and DL. When TTLM negotiation is performed, a TID can be mapped
to a subset of the established links, ranging from just one link to
all the links that have been set up.

A setup link is defined as enabled for non-AP MLD if at least one
TID is mapped to that link either in DL or UL. If no TIDs are mapped
in UL and DL for link, then that link is considered as disabled. At
any point in time, a TID shall always be mapped to at least one
setup link for both UL and DL. By default, all the setup links shall
be enabled.

This patch series covers the changes when a non default TTLM is
applied according to the procedure defined in IEEE P802.11be/D3.0
section 35.3.7.1.3 (Negotiation of TID-to-link mapping).

This patch set covers the below functionalities for negotiation of TTLM:

1. Support to send TTLM action frames - Request, Response and
Teardown from AP MLD.
2. Support to handle and parse the TTLM action frames received from
non-AP MLD.
3. Support to handle and parse association request containing TTLM
element from non-AP MLD.
4. Support to send association response from AP MLD for the received
TTLM association request.
5. Hostapd_cli interface support to send the TTLM action frames(request,
teardown) and override the response status code for association response.
6. User config to enable TTLM negotiation support in MLD capabilities
and operations field of basic multi-link element.

Patch-dependency : None

Feature: TTLM

Change-Id: I7fe93c6d053bd62b9ce025aca7a5e3337ced18d6
Signed-off-by: Dhanavandhana Kannan <dhanavan@qti.qualcomm.com>