UPSTREAM: PROTOCOL: hostapd: Add support for 6 GHz dynamic puncturing
Currently, if any of the 20 MHz subchannels in the operating
bandwidth of a 6 GHz AP interface is marked as disabled or no IR
for 6 GHz standard power (SP) mode, the channel is considered
unavailable for SP operation. This is suboptimal as we operate
in lower power modes even though such a channel can be used
for SP operation, if the disabled or no IR subchannel is
punctured.

Also, if any subchannel in the operating bandwidth has very low
SP EIRP based on the AFC payload, the EIRP of the entire
bandwidth gets reduced due to few subchannels having lower power.
The EIRP of the channel can be enhanced if the subchannel
with lower EIRP is punctured.

Add support for dynamic puncturing of 6 GHz standard power channels
in hostapd. This allows the AP to dynamically puncture the
subchannels that meet the criteria for dynamic puncturing:
 - The subchannel which does not support SP operation due to regulatory
    restrictions.
 - The subchannel which is marked as disabled or no IR
   for 6 GHz SP mode.
 - The subchannel which has very low EIRP based on the AFC
   payload.

Add support to further restrict the puncturing based on a configurable
EIRP threshold.

Add support to enable / disable dynamic puncturing using hostapd cli.

Add Add support to set the EIRP threshold for dynamic puncturing using
hostapd cli.

Fixes: None
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
n00-001-hostapd-add-support-for-6GHz-operation.patch
r32-001-hostapd-Add-support-to-receive-AFC-power-update-comp.patch
r72-a-UPSTREAM-PROTOCOL-hostapd-6-GHz-power-mode-channels-.patch
r72-d-UPSTREAM-PROTOCOL-hostapd-Clean-up-and-refactor-6-GH.patch
r72-e-UPSTREAM-PROTOCOL-hostapd-Support-for-SP-and-Punctur.patch
r72-h-UPSTREAM-PROTOCOL-hostapd-Apply-best-power-mode-for-.patch
r80-d-UPSTREAM-PROTOCOL-hostapd-Add-minimum-PSD-calculatio.patch
r81-UPSTREAM-PROTOCOL-hostapd-Use-OOBE-compliant-PSD-for.patch

Change-Id: Id4822cde776070742c23a64e09025bc1f0d4b4f1
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>