UPSTREAM: PROTOCOL: hostapd: Add hostapd_cli cmd for AFC payload clearing and AFC reset
Introduce a hostapd_cli command to trigger AFC payload clearing in
firmware. Notify the driver to request the firmware to clear the
stored AFC payload. The firmware clears the payload and sends an
event to switch to LPI or stop Tx. Subsequently, the AFC payload
stored in the driver is also cleared and it is notified to the
userspace. Apply this command only in the enterprise mode of AFC
operation.

Introduce a hostapd_cli command to trigger an AFC reset. Notify the
driver to request the firmware to reset the AFC expiry timer,
prompting an AFC Expiry event with start/renew as the subtype. This
forces the driver to restart AFC with a new request. Apply this
command only in the retail mode of AFC operation.

Fixes: None
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Path-dependency:
r32-001-hostapd-Add-support-to-receive-AFC-power-update-comp.patch

Change-Id: I97818f64d381870da233ffb77dce2c09c3ac3812
Signed-off-by: Amith A <amitajit@qti.qualcomm.com>