QSDK:PROTOCOL:HOSTAPD:Revert: Trigger scan when bss load update is enabled
This patch triggered an unnecessary AP scan when the AP interface was up
and beaconing. The AP interface was brought down because the scan failed
in the mac80211 layer, as the NL80211_SCAN_FLAG_AP flag was not set.
The patch was originally introduced to fix a hostapd crash caused by
scan_link being NULL; however, this issue is correctly handled upstream
in the following commit.

https://w1.fi/cgit/hostap/commit/?id=fd55dfe16e7a08c419e9430a67f5c121da6bb0e1

Tested with HS2.0 client. AP bringup successful and no hostapd crash was
seen

Change-Id: I6e4fcede374e987fcde7bed4a903a6460319fede
Signed-off-by: Sowmiya Sree Elavalagan <ssreeela@qti.qualcomm.com>