UPSTREAM: PROTOCOL: hostapd: Support 6 GHz SP mode with MLO
Currently, hostapd does not have the support for handling
an MLD with two 6 GHz links, where each link is in
different 6 GHz power modes.

All the AFC events from the driver are processed using the
iface corresponding to the first link which is part of the
MLD. This causes issues in the AFC event processing if there
are more than one links in an MLD if 6 GHz is not the first
link.

Also, when setting the 6 GHz power mode, the link id
is not passed to the driver, which causes issues in
the driver where there are multiple 6 GHz links in an MLD.

To fix this,
   (1) Process the AFC events from the driver for the
       iface corresponding to the hardware index for which
       the event is received.
   (2) Pass the link id as an attribute to the driver when
       setting the 6 GHz power mode for an MLD.
   (3) Process the AFC power mode change event based on the
       link id for which the power mode is changed.

Fixes: None
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
n00-001-hostapd-add-support-for-6GHz-operation.patch
r32-001-hostapd-Add-support-to-receive-AFC-power-update-comp.patch
r66-UPSTREAM-PROTOCOL-hostapd-Handle-AFC-Channel-Change.patch

Change-Id: Ia9ef1bdef5183e6a7d1cd39656689ac019bd68f1
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>