QSDK: PROTOCOL: hostapd: Remove patch no longer necessary
The changes in the patch q05-005-Handle-6G-beacon-report.patch are
no longer needed. There are four hunks present in the patch.

1) The first hunk adds duplicate 320 MHz opclass entry into global
opclass map. The duplication is due the entry already present in
upstream code which is merged during package upgrade.

2) The second and third hunk provides a fix for filling the supported op class in beacon report. But the fix is redundant as the beacon report is
correctly filled out and received, even without the fix.
This change checks HE Phy Capability flag for 80MHz/160MHz support
to fill the beacon report. But the check is redundant as the channel
flag is also checked for 80 MHz/160MHz support.

3) The fourth hunk has a check for 6 GHz when op class bw is 320 MHz,
which is unnecessary as there is only one 320MHz 6GHz opclass entry.

Change-Id: I6c4762dae442257b8a4ab1fdb0b1cf30e2583de5
Signed-off-by: Suraj P Kizhakkethil <quic_surapk@quicinc.com>