UPSTREAM:PROTOCOL: hostapd: Ignore BSS events from kernel when no active links are present
During network restart, hostapd crashes when processing BSS events
from the kernel. This is because hostapd tries to access a freed
BSS context.

While restarting the network, nl80211_stop_ap is received for last link,
the active link count drops to zero and BSS freed. As a result, the flink
is having a dangling pointer, and unexpected kernel event referencing that
BSS can lead to a segmentation fault.

This patch avoids the crash by ignoring BSS events when there are no
active links associated with the BSS.

Patch-dependency: None

Patch-work: None

Change-Id: I9708583462fb60fc2695c2503480076d5061e0f3
Signed-off-by: Karthik M <karm@qti.qualcomm.com>