UPSTREAM: PROTOCOL: hostapd: Set last octet of country string field to 0x04
In beacon frames, the last octet of the country string field in the
country IE is currently advertising a value of 0x20. However,
according to the WFA Agile Multiband Tech Spec, it should be set to
a value of 0x04. This adjustment is currently made only if the current
operating opclass is a 6 GHz opclass.

Hence set the last octet of the country string to 0x4, irrespective of the
current operating opclass.

Fixes: None
Patch-dependency: None
Patch-work: None

Change-Id: I07cff338a47f5f9bdcf45bb153ecd03470bda61f
Signed-off-by: Priyadarshnee Srinivasan <priys@qti.qualcomm.com>
(cherry picked from commit 9c43ee74bedf9d33e2fc52261551ff83d54a191c)