QSDK:HOSTAPD: Remove the 240MHz code from hostapd
240MHz is a QCN feature to support higher bandwidth in a 5GHz band.
This is an existing feature in QSDK hostapd and it cannot be upstreamed.
Remove this code from hostapd/patches and fix the conflicts in the further
patches to compile the hostapd.

The following patches have 240MHz related implementation. These patches are
deleted from the hostapd/patches folder.

1. q02-047-wpa_supplicant-Add-5GHz-240MHz-suppor.patch
2. q03-017-hostapd-240MHz-Q-Q-vendor-IE-support.patch
3. r04-008-hostapd-RNR-signalling-320MHz-BW-reduction-for-6GHz-.patch
4. r04-009-hostapd-Fix-VHT-center-frequency-for-EHT-240-MHz.patch
5. r42-QSDK-hostapd-fix-240-MHz-vendor-IE-for-AP-mode.patch
6. r96-QSDK-PRTOCOL-hostap-Add-240-MHz-vendor-element-in-as.patch

The below two files are removed from hostapd/patches folder.
1. q02-046-hostapd-Add-5GHz-240MHz-support.patch
2. q02-046-hostapd-DFS-ACS-5GHz-240MHz-support.patch

The above two patches contain implementation which is not related to 240MHz.
Add the below new patches, to handle this implementation related to channel
parameters in the ACS, DFS components:
1. q02-046-QSDK-hostapd-Add-API-to-verify-5GHz-frequency-range.patch
2. q02-046-QSDK-hostapd-Modify-ACS-and-DFS-implementation-for-c.patch

Signed-off-by: Sri Vidya Gunturi <sgunturi@qti.qualcomm.com>
Signed-off-by: Venkateswara Swamy Bandaru <vbandaru@qti.qualcomm.com>
Change-Id: Ia077ad7a1b436b486085a47f403cd5a07e5c29dc
(cherry picked from commit 87c2442cb08835155fe731a7d57237d4f8d45075)