QSDK: CP: hostapd: Fix dynamic chain mask during ACS
Prevent a NULL pointer crash in hostapd when the chain_mask command
is issued via hostapd_cli while the interface is in ACS and not yet
up.

Add a check to ensure the interface is up before processing the
command.

Also, handle nla_put_failure gracefully by freeing the netlink
message in failure scenarios.

Change-Id: I5b376b436733d1cbad1f443b5479b503abed3ef7
Signed-off-by: Rajat Soni <rajson@qti.qualcomm.com>