QSDK:PROTOCOL: hostap: fix ML association with raw ML elements
Currently, the ML element is parsed, and if an ML IE is found and the
driver supports the MLO flag, the authentication process continues.
Otherwise, a re-scan is triggered. There is a possibility that an ML IE is
present in the beacon frame—added manually via raw data in the
configuration—even though the MLO flag is not set. In such cases, ML
association consistently fails.

To address this issue, the parse ML IE only if the driver explicitly
supports MLO.

Patch Dependency: r38-wpa_s-initate-auth-when-all-configured.patch
Link: NA
Feature: MLO STA

Fixes: d6afddd78e25 ("wpa_supplicant: Initate auth for non EHT Station")
Signed-off-by: Aditya Kumar Singh <aditya.kumar.singh@oss.qualcomm.com>
Change-Id: I997068c67a15af5975894b12b132ab76d8fd5238