UPSTREAM:PROTOCOL: hostapd: properly deinit sm of non-ML sta connected to ML AP
Currently, when a non-ML client connects to an ML AP and the connection
is not fully established, ap_sta_disconnect_common() is triggered.
However, the state machine (sm) is not deinitialized because the AP is
an MLD and the assoc link ID does not match the connected link ID.
This results in sm being set to NULL, and later during ap_free_sta(),
the deinit function exits early due to sm being NULL, causing a memory leak.

Fix this issue by checking whether the connected client is  an ML client
instead of checking if the AP is an ML AP just like how it is done in
ap_free_sta() as well.

Patch Dependency: None
Link: [PATCH] hostapd: properly deinit sm of non-ML sta connected to ML AP
Feature: MLO general

Signed-off-by: Aditya Kumar Singh <aditya.kumar.singh@oss.qualcomm.com>
Change-Id: I58d56b2b04a0de4439b1a336128b316f99da45b5