UPSTREAM:HOSTAPD: Assign conf->opclass during CSA failure after DFS radar
When DFS radar is injected on an AP operating on a 5 GHz channel 52
in 20 MHz bandwidth, the AP goes down if the random target channel
selected for post-radar operation, is DFS.

The failure is observed in configured_fixed_chan_to_freq
(a sub-function of setup_interface2) as the opclass is incorrect for
the channel in which the AP is brought up.

This is because, in hostapd_dfs_request_channel_switch, when
hostapd_switch_channel returns a failure, the conf's channel number,
frequency, secondary channel, and bandwidth are set to those of the
parameters of the new target channel, but the conf->op_class is not
updated. Due to this, when configured_fixed_chan_to_freq is called as
part of hostapd_iface_enable, the conf->op_class does not match the
conf->channel.

To fix this issue, in hostapd_dfs_request_channel_switch,
set the conf->op_class to that of the new target channel.

Tag: DFS
Patch-dependency: r63-UPSTREAM-HOSTAPD-Update-conf-op_class-in-hostapd_cha.patch

Change-Id: Id732d0ab323189877348e96bb5e76d71e5813539
Signed-off-by: Hariharan Basuthkar <hbasuthk@qti.qualcomm.com>