UPSTREAM:PROTOCOL:CP:HOSTAPD: fix register to receive Robust AV SCS action frame
Currently during testing in testing_nl80211_register_frame() call, if
a Robust AV SCS action frame is already register it is returning the
error 'EALREADY' and due to this hw_sim test cases are failing.
But this is a success case only as already a action frame is
registered.

Hence, fix by return success in testing_nl80211_register_frame(),
if frame is already registered.

Fixes: I0e43d9e8ab2("UPSTREAM:PROTOCOL:CP:HOSTAPD Register to receive Robust AV SCS action frame")
Change-Id: Ifb093459c31b89994f3661887ffc6c0575366933
Signed-off-by: Sarika Sharma <sarishar@qti.qualcomm.com>