UPSTREAM:PROTOCOL:hostapd: Validate channel and bandwidth  during power mode configuration
When the AFC response includes only 160 MHz opclass CFIs and the
AP is configured for 320 MHz bandwidth, the AP correctly operates
in 320 MHz under LPI power mode, as no 320 MHz channels are
present in the AFC response.

However, if the power mode is manually switched to Standard Power
(SP) using the 'set_6ghz_power_mode' command via hostapd_cli, hostapd
continues to operate in 320 MHz under SP mode—despite the AFC response
not authorizing this configuration.

This issue stems from the kernel's SP channel list
construction logic, which marks a channel as SP-allowed if it is
supported by both the PSD and the 20 MHz CFI range. Since the 320 MHz range
satisfies both conditions, it is incorrectly flagged as SP-allowed. The
logic does not explicitly validate the channel CFIs against the
AFC response for the current configured bandwidth.

To address this, hostapd must validate that the channel is explicitly
listed in the AFC response before applying the SP power mode. This ensures
that only AFC-authorized channels are permitted for SP operation.

Fixes: 264ec03e8f2 (r28-hostapd-Add-hostapd-cli- support-to-switch-6GHZ-power.patch)
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
n00-001-hostapd-add-support-for-6GHz-operation.patch
r32-001-hostapd-Add-support-to-receive-AFC-power-update-comp.patch
r65-QSDK-HOSTAPD-Implement-Min-PSD-Calculation-and-Punct.patch
r72-a-UPSTREAM-PROTOCOL-hostapd-6-GHz-power-mode-channels-.patch
r72-d-UPSTREAM-PROTOCOL-hostapd-Clean-up-and-refactor-6-GH.patch
r72-e-UPSTREAM-PROTOCOL-hostapd-Support-for-SP-and-Punctur.patch
r28-hostapd-Add-hostapd-cli-support-to-switch-6GHZ-power.patch

Change-Id: I1a47bc82031ceb6fa9100f4e86e16c9b41753eb1
Signed-off-by: Priyadarshnee Srinivasan <priys@qti.qualcomm.com>